+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 * #4504: Bump gqlgen from v0.10.2 => v0.11.3
 * #4546: Refactor GraphQL purgeAddress to permit multiple target addresses
 * #4549: Implement GraphQL closeConnections for brokered and standard
+* #4510: Implement a GraphQL messaging endpoint API as a facade over addressspaces #4510
 
 ## 0.31.2
 * #4098: Added alert for high broker address memory usage